refactor: split app_protocol.rs into src/protocol/ submodules
4,036-line monolith split into:
- protocol/events.rs    — PlexiEvent (host→app)
- protocol/ui_nodes.rs  — UiNode + component tree types
- protocol/primitives.rs — shared structs (Rect, Modifiers, etc.)
- protocol/commands.rs  — RenderCommand, AppRequest, DrawCommand etc.
- protocol/mod.rs       — re-exports everything

src/app_protocol.rs becomes a 2-line shim (`pub use crate::protocol::*`)
so all existing callsites compile unchanged.

Pure refactor — zero behavioral change. 707/707 tests pass.

Closes #2044
